fix : new magnifique texture parsing

+++ bitmap.c
@@ -6,7 +6,7 @@
 /*   By: jtrauque <jtrauque@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/14 14:17:11 by jtrauque          #+#    #+#             */
-/*   Updated: 2021/01/16 17:35:20 by jtrauque         ###   ########.fr       */
+/*   Updated: 2021/01/19 19:30:27 by jtrauque         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,7 +69,7 @@ void	ft_create_bmp(t_pars *pars)
 
 int		ft_save_bmp(char *argv, t_pars *pars)
 {
-	size_t	n;
+	int	n;
 
 	if (ft_strlen("--save") != (n = ft_strlen(argv)))
 		return (ft_error("The second argument is not valid\n"));
